# openapi.yaml
openapi: 3.0.3
info:
  title: isc-fes API
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: ローカル開発環境
  - url: https://api.fes.iwasaki.ac.jp
    description: 本番環境
paths:
...
  /store-applications:
    get:
      operationId: getStoreApplications
      summary: 店舗申請一覧を取得する
      security:
        - AccountSession: []
      responses:
        "200":
          description: 店舗申請一覧
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetStoreApplicationsResponse"
        "401":
          description: 未ログイン
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 管理者権限がない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createStoreApplication
      summary: 店舗の作成申請を作成する
      security:
        - AccountSession: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
                - room
                - description
                - image
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                  example: たこ焼き屋

                room:
                  type: string
                  minLength: 1
                  maxLength: 50
                  example: 605

                description:
                  type: string
                  minLength: 1
                  maxLength: 1000
                  example: 外はカリカリ、中はトロトロのたこ焼きです。

                image:
                  type: string
                  format: binary
                  description: |
                    店舗画像。JPEG、PNG、WebPに対応する。
                    最大ファイルサイズは10MB。
      responses:
        "201":
          description: 店舗申請を作成した
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateStoreApplicationResponse"

        "400":
          description: リクエスト形式が不正
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "401":
          description: 未ログイン
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "413":
          description: アップロードされた画像が大きすぎる
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "415":
          description: 対応していない画像形式
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: 入力値または画像の内容が不正
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "503":
          description: 画像ストレージが一時的に利用できない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /store-applications/review-status/{store_id}:
    put:
      operationId: updateStoreApplicationReviewStatus
      summary: 店舗申請の審査ステータスを更新する
      security:
        - AccountSession: []
      parameters:
        - name: store_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reviewStatus
              properties:
                reviewStatus:
                  $ref: "#/components/schemas/StoreReviewStatus"
      responses:
        "200":
          description: 店舗申請の審査ステータスを更新した
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateStoreApplicationReviewStatusResponse"
        "400":
          description: リクエスト形式が不正
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未ログイン
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 管理者権限がない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 店舗申請が存在しない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: 更新できない審査ステータスへの変更 approved -> approved など
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    AccountSession:
      type: apiKey
      in: cookie
      name: isc_fes_account_session
  schemas:
...
    StoreReviewStatus:
      type: string
      enum:
        - pending
        - approved
        - rejected
    CreateStoreApplicationResponse:
      type: object
      required:
        - id
        - reviewStatus
        - submittedAt
      properties:
        id:
          type: string
          format: uuid
          description: |
            店舗申請のID。内部的にはstores.idと同一。

        reviewStatus:
          $ref: "#/components/schemas/StoreReviewStatus"

        submittedAt:
          type: string
          format: date-time
...
    Allergen:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
...
    UpdateStoreApplicationReviewStatusResponse:
      type: object
      required:
        - id
        - reviewStatus
      properties:
        id:
          type: string
          format: uuid
          description: |
            店舗申請のID。内部的にはstores.idと同一。

        reviewStatus:
          $ref: "#/components/schemas/StoreReviewStatus"
...
    GetStoreApplicationsResponse:
      type: object
      required:
        - data
        - total
      properties:
        total:
          type: integer
          description: 店舗申請の総数
        data:
          type: array
          items:
            $ref: "#/components/schemas/StoreApplication"
...
    StoreApplication:
      type: object
      required:
        - id
        - name
        - room
        - description
        - imageUrl
        - reviewStatus
        - submittedAt
        - allergens
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        room:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        reviewStatus:
          $ref: "#/components/schemas/StoreReviewStatus"
        submittedAt:
          type: string
          format: date-time
        allergens:
          type: array
          items:
            $ref: "#/components/schemas/Allergen"
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string